+++ app/models/activity.rb
@@ -21,6 +21,7 @@ class Activity < ApplicationRecord
     :country_step,
     :requires_additional_benefitting_countries_step,
     :intended_beneficiaries_step,
+    :gdi_step,
     :flow_step,
     :aid_type,
     :oda_eligibility_step,
@@ -42,6 +43,7 @@ class Activity < ApplicationRecord
   validates :recipient_country, presence: true, on: :country_step, if: :recipient_country?
   validates :requires_additional_benefitting_countries, inclusion: {in: [true, false]}, on: :requires_additional_benefitting_countries_step, if: :recipient_country?
   validates :intended_beneficiaries, presence: true, length: {maximum: 10}, on: :intended_beneficiaries_step, if: :requires_intended_beneficiaries?
+  validates :gdi, presence: true, on: :gdi_step
   validates :flow, presence: true, on: :flow_step
   validates :aid_type, presence: true, on: :aid_type_step
   validates :oda_eligibility, inclusion: {in: [true, false]}, on: :oda_eligibility_step